[feat] common폴더로 이동, 최근 24시간,30일에 대한 타입추가
- width, height, intervalType에 대한 props 추가
- 24시간, 30일로 나뉜것에 따른 x label함수 선언 , getDate 함수 선언
